Moved the validation classes
- Moved the validation classes, unnecessary request in namespace.
