+++ app/Http/Controllers/ItemTypeController.php
@@ -6,11 +6,11 @@ use App\Models\ItemType;
 use App\Option\Get;
 use App\Utilities\Header;
 use App\Utilities\Pagination as UtilityPagination;
-use App\Validators\Request\Route;
+use App\Validators\Route;
 use App\Models\Transformers\ItemType as ItemTypeTransformer;
 use App\Utilities\Response as UtilityResponse;
-use App\Validators\Request\SearchParameters;
-use App\Validators\Request\SortParameters;
+use App\Validators\SearchParameters;
+use App\Validators\SortParameters;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Config;
