+++ app/Http/Controllers/PermittedUserController.php
@@ -9,9 +9,9 @@ use App\Option\Post;
 use App\Utilities\Header;
 use App\Utilities\Pagination as UtilityPagination;
 use App\Utilities\RoutePermission;
-use App\Validators\Request\Route;
-use App\Validators\Request\SearchParameters;
-use App\Validators\Request\SortParameters;
+use App\Validators\Route;
+use App\Validators\SearchParameters;
+use App\Validators\SortParameters;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\Config;
