+++ app/Http/Controllers/ResourceTypeController.php
@@ -15,14 +15,14 @@ use App\Utilities\Header;
 use App\Utilities\Pagination as UtilityPagination;
 use App\Utilities\Request as UtilityRequest;
 use App\Utilities\RoutePermission;
-use App\Validators\Request\Parameters;
-use App\Validators\Request\Route;
+use App\Validators\Parameters;
+use App\Validators\Route;
 use App\Models\ResourceType;
 use App\Models\Transformers\ResourceType as ResourceTypeTransformer;
 use App\Utilities\Response as UtilityResponse;
-use App\Validators\Request\Fields\ResourceType as ResourceTypeValidator;
-use App\Validators\Request\SearchParameters;
-use App\Validators\Request\SortParameters;
+use App\Validators\Fields\ResourceType as ResourceTypeValidator;
+use App\Validators\SearchParameters;
+use App\Validators\SortParameters;
 use Exception;
 use Illuminate\Database\QueryException;
 use Illuminate\Http\JsonResponse;
