+++ app/Validators/Fields/RequestErrorLog.php
@@ -1,21 +1,22 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Validators\Request\Fields;
+namespace App\Validators\Fields;
 
-use App\Validators\Request\Fields\Validator as BaseValidator;
+use App\Validators\Fields\Validator as BaseValidator;
 use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Validator as ValidatorFacade;
 
 /**
- * Validation helper class for item category, returns the generated validator objects
+ * Validation helper class for request errors, returns the generated validator
+ * object
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ItemCategory extends BaseValidator
+class RequestErrorLog extends BaseValidator
 {
     /**
      * Return the validator object for the create request
@@ -26,24 +27,13 @@ class ItemCategory extends BaseValidator
      */
     public function create(array $options = []): Validator
     {
-        $decode = $this->hash->category()->decode(request()->input('category_id'));
-        $category_id = null;
-        if (count($decode) === 1) {
-            $category_id = $decode[0];
-        }
-
         return ValidatorFacade::make(
-            ['category_id' => $category_id],
-            Config::get('api.item-category.validation.POST.fields'),
-            $this->translateMessages('api.item-category.validation.POST.messages')
+            request()->all(),
+            Config::get('api.request-error-log.validation.POST.fields'),
+            $this->translateMessages('api.request-error-log.validation.POST.messages')
         );
     }
 
-    /**
-     * @param array $options
-     *
-     * @return Validator
-     */
     public function update(array $options = []): Validator
     {
         // TODO: Implement update() method.
