+++ app/Validators/Fields/Resource.php
@@ -1,31 +1,30 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Validators\Request\Fields;
+namespace App\Validators\Fields;
 
-use App\Rules\ResourceTypeName;
-use App\Validators\Request\Fields\Validator as BaseValidator;
+use App\Validators\Fields\Validator as BaseValidator;
 use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Validator as ValidatorFacade;
 
 /**
- * Validation helper class for resource types, returns the generated validator objects
+ * Validation helper class for resources, returns the generated validator objects
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ResourceType extends BaseValidator
+class Resource extends BaseValidator
 {
     /**
      * Create the validation rules for the create request
      *
-     * @param integer $user_id
+     * @param integer $resource_type_id
      *
      * @return array
      */
-    private function createRules(int $user_id): array
+    private function createRules(int $resource_type_id): array
     {
         return array_merge(
             [
@@ -33,14 +32,10 @@ class ResourceType extends BaseValidator
                     'required',
                     'string',
                     'max:255',
-                    new ResourceTypeName($user_id)
+                    'unique:resource,name,null,id,resource_type_id,' . $resource_type_id
                 ],
-                'item_type_id' => [
-                    'required',
-                    'exists:item_type,id'
-                ]
             ],
-            Config::get('api.resource-type.validation.POST.fields')
+            Config::get('api.resource.validation.POST.fields')
         );
     }
 
@@ -48,11 +43,11 @@ class ResourceType extends BaseValidator
      * Create the validation rules for the update request
      *
      * @param integer $resource_type_id
-     * @param integer $user_id
+     * @param integer $resource_id
      *
      * @return array
      */
-    private function updateRules(int $resource_type_id, int $user_id): array
+    private function updateRules(int $resource_type_id, int $resource_id): array
     {
         return array_merge(
             [
@@ -60,10 +55,10 @@ class ResourceType extends BaseValidator
                     'sometimes',
                     'string',
                     'max:255',
-                    new ResourceTypeName($user_id, $resource_type_id)
-                ]
+                    'unique:resource,name,'. $resource_id . ',id,resource_type_id,' . $resource_type_id
+                ],
             ],
-            Config::get('api.resource-type.validation.PATCH.fields')
+            Config::get('api.resource.validation.PATCH.fields')
         );
     }
 
@@ -81,33 +76,25 @@ class ResourceType extends BaseValidator
      */
     public function create(array $options = []): Validator
     {
-        $decode = $this->hash->itemType()->decode(request()->input('item_type_id'));
-        $item_type_id = null;
-        if (count($decode) === 1) {
-            $item_type_id = $decode[0];
-        }
+        $this->requiredIndexes(['resource_type_id'], $options);
 
         return ValidatorFacade::make(
-            array_merge(
-                request()->all(),
-                ['item_type_id' => $item_type_id]
-            ),
-            $this->createRules($options['user_id']),
-            $this->translateMessages('api.resource-type.validation.POST.messages')
+            request()->all(),
+            self::createRules(intval($options['resource_type_id'])),
+            $this->translateMessages('api.resource.validation.POST.messages')
         );
     }
 
     /**
      * @param array $options
-     *
      * @return Validator
      */
     public function update(array $options = []): Validator
     {
         return ValidatorFacade::make(
             request()->all(),
-            $this->updateRules($options['resource_type_id'], $options['user_id']),
-            $this->translateMessages('api.resource-type.validation.PATCH.messages')
+            $this->updateRules($options['resource_type_id'], $options['resource_id']),
+            $this->translateMessages('api.resource.validation.PATCH.messages')
         );
     }
 }
