+++ app/Validators/Routes/Category.php
@@ -1,37 +1,40 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Validators\Request\Routes;
+namespace App\Validators\Routes;
 
 use App\Models\ResourceTypeAccess;
 
 /**
- * Validate the route params to a resource type
+ * Validate the route params to a category
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ResourceType
+class Category
 {
     /**
      * Validate that the user is able to view the requested resource type based
      * on their permitted resource types, needs to be in their group or public
      *
-     * @param string|int $resource_type_id
+     * @param int $resource_type_id
+     * @param int $category_id
      * @param array $permitted_resource_types
      *
      * @return boolean
      */
     static public function existsToUserForViewing(
         $resource_type_id,
+        $category_id,
         array $permitted_resource_types
     ): bool
     {
         if (
-            $resource_type_id === 'nill' ||
-            (new ResourceTypeAccess())->resourceTypeExistsToUser(
+            $category_id === 'nill' ||
+            (new ResourceTypeAccess())->categoryExistsToUser(
                 $resource_type_id,
+                $category_id,
                 $permitted_resource_types
             ) === false
         ) {
@@ -45,20 +48,23 @@ class ResourceType
      * Validate that the user is able to manage the requested resource type
      * based on their permitted resource types, needs to be in their group
      *
-     * @param string|int $resource_type_id
+     * @param int $resource_type_id
+     * @param int $category_id
      * @param array $permitted_resource_types
      *
      * @return boolean
      */
     static public function existsToUserForManagement(
         $resource_type_id,
+        $category_id,
         array $permitted_resource_types
     ): bool
     {
         if (
-            $resource_type_id === 'nill' ||
-            (new ResourceTypeAccess())->resourceTypeExistsToUser(
+            $category_id === 'nill' ||
+            (new ResourceTypeAccess())->categoryExistsToUser(
                 $resource_type_id,
+                $category_id,
                 $permitted_resource_types,
                 true
             ) === false
