+++ app/Validators/Routes/Item.php
@@ -1,27 +1,26 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Validators\Request\Routes;
+namespace App\Validators\Routes;
 
 use App\Models\ResourceTypeAccess;
 
 /**
- * Validate the route params to an item category
+ * Validate the route params to an item
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ItemCategory
+class Item
 {
     /**
-     * Validate that the user is able to view the requested item category based
+     * Validate that the user is able to view the requested item based
      * on their permitted resource types, needs to be in their group or public
      *
      * @param string|int $resource_type_id
      * @param string|int $resource_id
      * @param string|int $item_id
-     * @param string|int $item_category_id
      * @param array $permitted_resource_types
      *
      * @return boolean
@@ -30,7 +29,6 @@ class ItemCategory
         $resource_type_id,
         $resource_id,
         $item_id,
-        $item_category_id,
         array $permitted_resource_types
     ): bool
     {
@@ -38,12 +36,10 @@ class ItemCategory
             $resource_type_id === 'nill' ||
             $resource_id === 'nill' ||
             $item_id === 'nill' ||
-            $item_category_id === 'nill' ||
-            (new ResourceTypeAccess())->itemCategoryExistsToUser(
+            (new ResourceTypeAccess())->itemExistsToUser(
                 $resource_type_id,
                 $resource_id,
                 $item_id,
-                $item_category_id,
                 $permitted_resource_types
             ) === false
         ) {
@@ -54,13 +50,12 @@ class ItemCategory
     }
 
     /**
-     * Validate that the user is able to manage the requested item category
+     * Validate that the user is able to manage the requested item
      * based on their permitted resource types, needs to be in their group
      *
      * @param string|int $resource_type_id
      * @param string|int $resource_id
      * @param string|int $item_id
-     * @param string|int $item_category_id
      * @param array $permitted_resource_types
      *
      * @return boolean
@@ -69,7 +64,6 @@ class ItemCategory
         $resource_type_id,
         $resource_id,
         $item_id,
-        $item_category_id,
         array $permitted_resource_types
     ): bool
     {
@@ -77,12 +71,10 @@ class ItemCategory
             $resource_type_id === 'nill' ||
             $resource_id === 'nill' ||
             $item_id === 'nill' ||
-            $item_category_id === 'nill' ||
-            (new ResourceTypeAccess())->itemCategoryExistsToUser(
+            (new ResourceTypeAccess())->itemExistsToUser(
                 $resource_type_id,
                 $resource_id,
                 $item_id,
-                $item_category_id,
                 $permitted_resource_types,
                 true
             ) === false
