+++ app/Validators/Routes/ItemCategory.php
@@ -1,25 +1,27 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Validators\Request\Routes;
+namespace App\Validators\Routes;
 
 use App\Models\ResourceTypeAccess;
 
 /**
- * Validate the route params to a resource
+ * Validate the route params to an item category
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class Resource
+class ItemCategory
 {
     /**
-     * Validate that the user is able to view the requested resource based
+     * Validate that the user is able to view the requested item category based
      * on their permitted resource types, needs to be in their group or public
      *
      * @param string|int $resource_type_id
      * @param string|int $resource_id
+     * @param string|int $item_id
+     * @param string|int $item_category_id
      * @param array $permitted_resource_types
      *
      * @return boolean
@@ -27,15 +29,21 @@ class Resource
     static public function existsToUserForViewing(
         $resource_type_id,
         $resource_id,
+        $item_id,
+        $item_category_id,
         array $permitted_resource_types
     ): bool
     {
         if (
             $resource_type_id === 'nill' ||
             $resource_id === 'nill' ||
-            (new ResourceTypeAccess())->resourceExistsToUser(
-                $resource_id,
+            $item_id === 'nill' ||
+            $item_category_id === 'nill' ||
+            (new ResourceTypeAccess())->itemCategoryExistsToUser(
                 $resource_type_id,
+                $resource_id,
+                $item_id,
+                $item_category_id,
                 $permitted_resource_types
             ) === false
         ) {
@@ -46,11 +54,13 @@ class Resource
     }
 
     /**
-     * Validate that the user is able to manage the requested resource
+     * Validate that the user is able to manage the requested item category
      * based on their permitted resource types, needs to be in their group
      *
      * @param string|int $resource_type_id
      * @param string|int $resource_id
+     * @param string|int $item_id
+     * @param string|int $item_category_id
      * @param array $permitted_resource_types
      *
      * @return boolean
@@ -58,15 +68,21 @@ class Resource
     static public function existsToUserForManagement(
         $resource_type_id,
         $resource_id,
+        $item_id,
+        $item_category_id,
         array $permitted_resource_types
     ): bool
     {
         if (
             $resource_type_id === 'nill' ||
             $resource_id === 'nill' ||
-            (new ResourceTypeAccess())->resourceExistsToUser(
-                $resource_id,
+            $item_id === 'nill' ||
+            $item_category_id === 'nill' ||
+            (new ResourceTypeAccess())->itemCategoryExistsToUser(
                 $resource_type_id,
+                $resource_id,
+                $item_id,
+                $item_category_id,
                 $permitted_resource_types,
                 true
             ) === false
