+++ app/Validators/Routes/Resource.php
@@ -1,26 +1,25 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Validators\Request\Routes;
+namespace App\Validators\Routes;
 
 use App\Models\ResourceTypeAccess;
 
 /**
- * Validate the route params to an item
+ * Validate the route params to a resource
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class Item
+class Resource
 {
     /**
-     * Validate that the user is able to view the requested item based
+     * Validate that the user is able to view the requested resource based
      * on their permitted resource types, needs to be in their group or public
      *
      * @param string|int $resource_type_id
      * @param string|int $resource_id
-     * @param string|int $item_id
      * @param array $permitted_resource_types
      *
      * @return boolean
@@ -28,18 +27,15 @@ class Item
     static public function existsToUserForViewing(
         $resource_type_id,
         $resource_id,
-        $item_id,
         array $permitted_resource_types
     ): bool
     {
         if (
             $resource_type_id === 'nill' ||
             $resource_id === 'nill' ||
-            $item_id === 'nill' ||
-            (new ResourceTypeAccess())->itemExistsToUser(
-                $resource_type_id,
+            (new ResourceTypeAccess())->resourceExistsToUser(
                 $resource_id,
-                $item_id,
+                $resource_type_id,
                 $permitted_resource_types
             ) === false
         ) {
@@ -50,12 +46,11 @@ class Item
     }
 
     /**
-     * Validate that the user is able to manage the requested item
+     * Validate that the user is able to manage the requested resource
      * based on their permitted resource types, needs to be in their group
      *
      * @param string|int $resource_type_id
      * @param string|int $resource_id
-     * @param string|int $item_id
      * @param array $permitted_resource_types
      *
      * @return boolean
@@ -63,18 +58,15 @@ class Item
     static public function existsToUserForManagement(
         $resource_type_id,
         $resource_id,
-        $item_id,
         array $permitted_resource_types
     ): bool
     {
         if (
             $resource_type_id === 'nill' ||
             $resource_id === 'nill' ||
-            $item_id === 'nill' ||
-            (new ResourceTypeAccess())->itemExistsToUser(
-                $resource_type_id,
+            (new ResourceTypeAccess())->resourceExistsToUser(
                 $resource_id,
-                $item_id,
+                $resource_type_id,
                 $permitted_resource_types,
                 true
             ) === false
