+++ app/Validators/Routes/ResourceType.php
@@ -1,45 +1,37 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Validators\Request\Routes;
+namespace App\Validators\Routes;
 
 use App\Models\ResourceTypeAccess;
 
 /**
- * Validate the route params to a sub category
+ * Validate the route params to a resource type
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018-2020
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class Subcategory
+class ResourceType
 {
     /**
-     * Validate that the user is able to view the requested subcategory based
+     * Validate that the user is able to view the requested resource type based
      * on their permitted resource types, needs to be in their group or public
      *
-     * @param integer $resource_type_id
-     * @param integer $category_id
-     * @param integer $subcategory_id
+     * @param string|int $resource_type_id
      * @param array $permitted_resource_types
      *
      * @return boolean
      */
     static public function existsToUserForViewing(
-        int $resource_type_id,
-        int $category_id,
-        int $subcategory_id,
+        $resource_type_id,
         array $permitted_resource_types
     ): bool
     {
         if (
             $resource_type_id === 'nill' ||
-            $category_id === 'nill' ||
-            $subcategory_id === 'nill' ||
-            (new ResourceTypeAccess())->subcategoryExistsToUser(
+            (new ResourceTypeAccess())->resourceTypeExistsToUser(
                 $resource_type_id,
-                $category_id,
-                $subcategory_id,
                 $permitted_resource_types
             ) === false
         ) {
@@ -50,31 +42,23 @@ class Subcategory
     }
 
     /**
-     * Validate that the user is able to manage the requested subcategor
+     * Validate that the user is able to manage the requested resource type
      * based on their permitted resource types, needs to be in their group
      *
-     * @param integer $resource_type_id
-     * @param integer $category_id
-     * @param integer $subcategory_id
+     * @param string|int $resource_type_id
      * @param array $permitted_resource_types
      *
      * @return boolean
      */
     static public function existsToUserForManagement(
-        int $resource_type_id,
-        int $category_id,
-        int $subcategory_id,
+        $resource_type_id,
         array $permitted_resource_types
     ): bool
     {
         if (
             $resource_type_id === 'nill' ||
-            $category_id === 'nill' ||
-            $subcategory_id === 'nill' ||
-            (new ResourceTypeAccess())->subcategoryExistsToUser(
+            (new ResourceTypeAccess())->resourceTypeExistsToUser(
                 $resource_type_id,
-                $category_id,
-                $subcategory_id,
                 $permitted_resource_types,
                 true
             ) === false
