+++ app/Validators/SortParameters.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace App\Validators\Request;
+namespace App\Validators;
 
 /**
  * Fetch and validate any sort parameters
